Add Feature: Add Partial Data Collection Mode
When the "Ready" button is pressed, a dialog will pop up for the user to input the experiment setup.

The dialog will also display the estimated experiment time based on the user's input.

After the user clicks the "Finish" button in the dialog:

The window will freeze until the experiment finishes or until the user restarts the app to prevent unintended touch errors.
A folder structure will be created to store the sent and received data, following the structure shared in the Google Drive and WeChat group.